backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.11.0
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.119.0
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.45.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.1
langchain-core==1.0.0
langchain-google-genai==3.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
mcp==1.18.0
multidict==6.7.0
nltk==3.9.2
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pillow==12.0.0
praw==7.8.1
prawcore==2.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyparsing==3.2.5
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tavily-python==0.7.12
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vaderSentiment==3.3.2
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0